Replace Codeception with PSR-4 phpunit tests
