--- codeception.yml
@@ -1,25 +1,0 @@
-actor: Tester
-paths:
-    tests: tests
-    log: tests/_output
-    data: tests/_data
-    support: tests/_support
-    envs: tests/_envs
-settings:
-    bootstrap: _bootstrap.php
-    colors: true
-    memory_limit: 1024M
-coverage:
-    enabled: true
-    include:
-        - src/*
-extensions:
-    enabled:
-        - Codeception\Extension\RunFailed
-modules:
-    config:
-        Db:
-            dsn: ''
-            user: ''
-            password: ''
-            dump: tests/_data/dump.sql
