--- tests/_bootstrap.php
@@ -1,8 +1,0 @@
-<?php
-
-// Codeception Specify deep-clones everything by default...
-ini_set('xdebug.max_nesting_level', 1000);
-
-date_default_timezone_set('UTC');
-
-require __DIR__ . '/../vendor/autoload.php';
