--- tests/_data/dump.sql
@@ -1,1 +1,0 @@
-/* Replace this file with actual dump of your database */
