--- tests/_support/Helper/Unit.php
@@ -1,9 +1,0 @@
-<?php
-namespace Helper;
-
-// here you can define custom actions
-// all public methods declared in helper class will be available in $I
-
-class Unit extends \Codeception\Module
-{
-}
