--- tests/unit/_bootstrap.php
@@ -1,4 +1,0 @@
-<?php
-// Here you can initialize variables that will be available to your tests
-
-\Codeception\Specify\Config::setDeepClone(false);
